Rename package XdgMenu to Menu, rename XdgMenuWidget to MenuWidget.
