+++ src/System/Taffybar/Menu/COPYING
