+++ taffybar.cabal
@@ -102,10 +102,9 @@ library
                    System.Information.CPU,
                    System.Information.CPU2,
                    System.Information.DiskIO
-                   System.Taffybar.XdgMenu.DesktopEntry,
-                   System.Taffybar.XdgMenu.DesktopEntryCondition,
-                   System.Taffybar.XdgMenu.XdgMenu,
-                   System.Taffybar.XdgMenu.XdgMenuWidget
+                   System.Taffybar.Menu.DesktopEntry,
+                   System.Taffybar.Menu.XdgMenu,
+                   System.Taffybar.Menu.MenuWidget
   other-modules: System.Taffybar.StrutProperties,
                  Paths_taffybar
